requirements.txt
asgiref==3.7.2
astroid==2.15.6
backports.zoneinfo==0.2.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
dill==0.3.6
Django==4.2.9
django-cors-headers==4.3.1
Flask==2.2.2
Flask-Cors==4.0.0
flask-http-middleware==0.2.1
idna==3.4
importlib-metadata==6.7.0
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
lazy-object-proxy==1.9.0
MarkupSafe==2.1.3
mccabe==0.7.0
MetaTrader5==5.0.4424
numpy==1.24.4
peewee==3.15.4
platformdirs==3.9.1
pycodestyle==2.10.0
PyJWT==2.6.0
pylint==2.17.4
PyMySQL==1.0.2
python-dotenv==0.21.0
python-json-logger==2.0.7
requests==2.28.2
sqlparse==0.4.4
stripe==5.4.0
tomli==2.0.1
tomlkit==0.11.8
typing_extensions==4.7.1
tzdata==2023.4
urllib3==1.26.16
Werkzeug==2.3.6
wrapt==1.15.0
zipp==3.15.0
